requirements.txt
openai==0.28.0
openai_whisper==20240930
python-dotenv==1.0.1
tkinterdnd2==0.3.0
torch==2.0.1